% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMstep2.R
\name{CMstep2}
\alias{CMstep2}
\title{Conditional weighted ML estimation of theta given mu and Sigma.}
\usage{
CMstep2(
  X,
  weights = NULL,
  mu,
  Sigma,
  formula = "direct",
  naMat = NULL,
  obsInd = NULL,
  misInd = NULL
)
}
\arguments{
\item{X}{Data matrix.}

\item{weights}{Vector of weights.}

\item{mu}{Mean vector.}

\item{Sigma}{Scalar matrix.}

\item{formula}{Method used to calculate the density: \code{"direct"}, \code{"indirect"}, \code{"series"}.}

\item{naMat}{Matrix of missing value indicators for data matrix.}

\item{obsInd}{Indices of rows containing complete observations.}

\item{misInd}{Indices of rows containing missing observations.}
}
\value{
\item{theta}{Estimated inflation parameter.}
}
\description{
Conditional weighted ML estimation of theta given mu and Sigma.
}
\examples{
d <- 2
mu <- rep(0,d)
Sigma <- diag(d)
set.seed(5)
X <- rmtin(n=100,mu,Sigma,theta=0.4)$X
CMstep2(X=X,mu=mu,Sigma=Sigma)

}
