% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmtin.R
\name{dmtin}
\alias{dmtin}
\title{Density of a MTIN distribution.}
\usage{
dmtin(x, mu = rep(0, d), Sigma, theta = 0.01, formula = "direct")
}
\arguments{
\item{x}{Matrix.}

\item{mu}{Mean vector.}

\item{Sigma}{Scale matrix.}

\item{theta}{Inflation parameter.}

\item{formula}{Method used to calculate the density: \code{"direct"}, \code{"indirect"}, \code{"series"}.}
}
\value{
The value(s) of the density in \code{x}.
}
\description{
Density of a MTIN distribution.
}
\examples{
d <- 3
x <- matrix(rnorm(d*2),2,d)
dmtin(x,mu=rep(0,d),Sigma=diag(d),theta=0.4,formula="indirect")

}
