% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_clusters.R
\name{init_clusters}
\alias{init_clusters}
\title{Initialize membership labels}
\usage{
init_clusters(
  X,
  G,
  obsInd = NULL,
  init_method = c("mclust", "kmedoids", "kmeans", "heirarchical")
)
}
\arguments{
\item{G}{An integer representing the number of clusters.}

\item{obsInd}{Indices of rows containing complete observations.}

\item{init_method}{Model used for initialization: \code{'mclust'}, \code{"kmedoids"}, \code{"kmeans"}, \code{"heirarchical"}}

\item{x}{A matrix whose columns are variables.}

\item{labels}{Vector of membership labels to be included if using manual initialization.}
}
\value{
\item{Z}{An (n x G) matrix representing membership labels.}
}
\description{
Initialize membership labels
}
\keyword{internal}
