% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmtin.R
\name{rmtin}
\alias{rmtin}
\title{Generate random numbers from the MTIN Distribution.}
\usage{
rmtin(n, mu = rep(0, d), Sigma, theta = 0.01, norm.dens = "dmnorm")
}
\arguments{
\item{n}{Number of points to be generated.}

\item{mu}{Mean vector.}

\item{Sigma}{Scale matrix.}

\item{theta}{Inflation parameter.}

\item{norm.dens}{Method used to calculate the density of the normal distribution: \code{"dmnorm"} (see \code{\link[mnormt]{dmnorm}}) or \code{"dmvnorm"} (see \code{\link[mvtnorm]{Mvnorm}}).}
}
\value{
A list with the following elements:
\item{X}{Data matrix.}
\item{w}{Vector of weights.}
}
\description{
Generate random numbers from the MTIN Distribution.
}
\examples{
d <- 3
rmtin(10,mu=rep(0,d),Sigma=diag(d),theta=0.4)

}
