% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinreg.R
\name{tinreg}
\alias{tinreg}
\title{Fit cluster weighted regression model using the MTIN distribution}
\usage{
tinreg(
  x,
  y,
  G = 1,
  max_iter = 100,
  tol = 10^-1,
  model = c("fixed", "random"),
  formula = c("direct", "indirect", "series"),
  init_method = c("mclust", "kmedoids", "kmeans", "heirarchical"),
  verbose = FALSE,
  plot_likelihood = FALSE
)
}
\arguments{
\item{x}{A matrix whose columns are the predictor variables.}

\item{y}{A matrix whose columns are the response variables.}

\item{G}{An integer representing the number of clusters.}

\item{max_iter}{An integer representing the maximum number of iterations.}

\item{tol}{Float representing tolerance for Aitken's acceleration criterion.}

\item{model}{Type of regression model to fit: \code{"fixed"}, \code{"random"}}

\item{formula}{Method used to calculate density passed to dmtin: \code{"direct"}, \code{"indirect"}, \code{"series"}}

\item{init_method}{Model used for initialization: \code{'mclust'}, \code{"kmedoids"}, \code{"kmeans"}, \code{"heirarchical"}}
}
\value{
A list with the following elements:
\item{Beta}{List of regression coefficients.}
\item{Mu_x}{List of mean vectors for independent variable.}
\item{Sigma_x}{List of scale matrices for independent variable.}
\item{Theta_x}{List of tailedness parameters for independent variable.}
\item{Sigma_y}{List of scale matrices for dependent variable.}
\item{Theta_y}{List of tailedness parameters for dependent variable.}
\item{Pi}{Mixing proportions.}
\item{Z}{Expected value of Z.}
\item{W_y}{Expected value of W for dependent variable.}
\item{W_X}{Expected value of W for independent variable.}
\item{cluster}{Membership labels.}
\item{L}{List of log-likelihoods at each iteration.}
\item{Model}{Type of model fitted.}
\item{AIC}{Akaike information criterion.}
\item{BIC}{Bayesian information criterion.}
\item{KIC}{Kullback information criterion.}??
}
\description{
Fit cluster weighted regression model using the MTIN distribution
}
\examples{
\donttest{
data <- palmerpenguins::penguins
keep <- !rowSums(is.na(data[,3:6])) == 4
data <- data[keep,]

res <- tinreg(x = data$flipper_length_mm, y = data$bill_length_mm, G = 3, model = 'random')
plot(res$X[,2], res$Y, col = res$cluster)
abline(a = res$Beta[[1]][1,1], b = res$Beta[[1]][2,1], col = 'black')
abline(a = res$Beta[[2]][1,1], b = res$Beta[[2]][2,1], col = 'red')
abline(a = res$Beta[[3]][1,1], b = res$Beta[[3]][2,1], col = 'green')
}

}
