% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinclust.R
\name{tinclust}
\alias{tinclust}
\title{Fit cluster weighted regression model using the MTIN distribution}
\usage{
tinclust(x)
}
\arguments{
\item{x}{A data matrix.}

\item{G}{An integer representing the number of clusters.}

\item{max_iter}{An integer representing the maximum number of iterations.}

\item{tol}{Float representing tolerance for Aitken's acceleration criterion.}

\item{formula}{Method used to calculate density passed to \code{dmtin}: \code{"direct"}, \code{"indirect"}, \code{"series"}}

\item{init_method}{Model used for initialization: \code{'mclust'}, \code{"kmedoids"}, \code{"kmeans"}, \code{"heirarchical"}}
}
\value{
A list with the following elements:
\item{Beta}{List of regression coefficients.}
\item{Mu_x}{List of mean vectors for independent variable.}
\item{Sigma_x}{List of scale matrices for independent variable.}
\item{Theta_x}{List of tailedness parameters for independent variable.}
\item{Sigma_y}{List of scale matrices for dependent variable.}
\item{Theta_y}{List of tailedness parameters for dependent variable.}
\item{Pi}{Mixing proportions.}
\item{Z}{Expected value of Z.}
\item{W_y}{Expected value of W for dependent variable.}
\item{W_X}{Expected value of W for independent variable.}
\item{cluster}{Membership labels.}
\item{L}{List of log-likelihoods at each iteration.}
\item{Model}{Type of model fitted.}
\item{AIC}{Akaike information criterion.}
\item{BIC}{Bayesian information criterion.}
\item{KIC}{Kullback information criterion.}??
}
\description{
Fit cluster weighted regression model using the MTIN distribution
}
